+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 - `TcpDialer` now simply takes a `Url` parameter, rather than a host and a port or a socket address.
 - Added a convenience `repo.dial_tcp()` method to simplify construction of `TcpDialer`. 
 - Allow documents syncing over the TCP transport to be up to 8gb size instead of Tokio's default 8mb frame size
+- Exposed receiving `ConnectionHandle`s via `accept()`. Users can now subscribe to an `events()` stream directly
+  on the handle, or `await` for `handshake_completed()`.
 
 ## 0.8.0 - 2024-03-06